temp/requirements.txt
boto3
pillow
pydantic